# Conflicts:
#	src/test/java/appointmentDAOTest/DeleteAppointmentServiceTest.java

+++ src/main/java/com/bnta/patient/PatientService.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
+import java.util.Objects;
+import java.util.Optional;
 import java.util.List;
 
 @Service
@@ -17,25 +19,31 @@ public class PatientService {
         this.patientDAO = patientDAO;
     }
 
-    public int addNewPatient(Patient patient){
-        if(patient == null){
+    public boolean doesPatientWithIdExist(Integer patientNhsId) {
+        return patientDAO
+                .selectAllPatients()
+                .stream()
+                .anyMatch(p -> Objects.equals(p.getPatientNhsId(), patientNhsId));
+    }
+
+    public int addNewPatient(Patient patient) {
+        if (patient == null) {
             throw new IllegalArgumentException("Patient cannot be null");
         }
-        if (patient.getPatientNhsId() == 0 ||
-                patient.getPatientName() == null ||
+        if (patient.getPatientName() == null ||
                 patient.getPatientEmailAddress() == null ||
                 patient.getPatientPhoneNumber() == null ||
-                patient.getBloodType().toString() == null) {
+                patient.getBloodType() == null) {
             throw new IllegalStateException("Patient cannot have empty fields");
         }
 
-     //   boolean exists = doesPersonWithIdExists(person.getId());
-     //   if (exists) {
-        //     throw new IllegalStateException("person with id " + person.getId() + " already exists");
-      //  }
+        boolean exists = doesPatientWithIdExist(patient.getPatientNhsId());
+        if (exists) {
+            throw new IllegalStateException("patient with id " + patient.getPatientNhsId() + " already exists");
+        }
 
-     //   int result = personDAO.insertPerson(person);
-      //  return result;
+  /*      int result = patientDAO.insertPatient(patient);
+        return result;*/
 
         int result = patientDAO.insertPatient(patient);
         if (result != 1) {
@@ -45,15 +53,19 @@ public class PatientService {
         }
     }
 
+// to test ? we incorporated it to others
     public Patient findPatientById(Integer id) {
-        try {
-            Patient output = patientDAO.selectPatientById(id);
-            return output;
-        } catch (EmptyResultDataAccessException e) {
-            throw new PatientNotFoundException("Patient with ID '" + id + "' not found.");
+        boolean exists = doesPatientWithIdExist(id);
+        if (exists) {
+            try {
+                return patientDAO.selectPatientById(id);
+            } catch (EmptyResultDataAccessException e) {
+                throw new PatientNotFoundException("Patient with ID '" + id + "' not found.");
+            }
         }
+        return patientDAO.selectPatientById(id);
     }
-
+// need to test
     public List<Patient> findAllPatients() {
         try {
             return patientDAO.selectAllPatients();
@@ -62,27 +74,33 @@ public class PatientService {
         }
     }
 
-    public void updatePatient(Integer id, Patient update) {
-        try {
-            int result = patientDAO.updatePatient(id, update);
-            if (result != 1) {
-                throw new IllegalStateException("Could not update car.");
+    public int updatePatient(Integer id, Patient update) {
+        if (update == null) {
+            throw new IllegalArgumentException("Patient cannot be null");
+        } else if (update.getPatientName() == null ||
+                update.getPatientEmailAddress() == null ||
+                update.getPatientPhoneNumber() == null ||
+                update.getBloodType() == null) {
+            throw new IllegalStateException("Patient cannot have empty fields");
+        } else {
+            try {
+                return patientDAO.updatePatient(id, update);
+
+            } catch (EmptyResultDataAccessException e) {
+                throw new PatientNotFoundException("Patient with ID '" + id + "' not found.");
             }
-        } catch (EmptyResultDataAccessException e) {
+        }
+    }
+
+    public int deletePatientById(Integer id) {
+        boolean exists = doesPatientWithIdExist(id);
+        if (exists) {
+            Patient delPatient = patientDAO.selectPatientById(id);
+            return patientDAO.deletePatient(delPatient);
+        } else {
             throw new PatientNotFoundException("Patient with ID '" + id + "' not found.");
         }
     }
 
-    public void deletePatientById(Integer id) {
-        try {
-            Patient delPatient = patientDAO.selectPatientById(id);
-            int result = patientDAO.deletePatient(delPatient);
-            if (result != 1) {
-                throw new IllegalStateException("Could not delete patient.");
-            }
-        } catch (EmptyResultDataAccessException e) {
-            throw new PatientNotFoundException("Patient with ID '" + id + "' not found.");
-        }
-    }
 }
